requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.4.0
babel==2.8.0              # via flask-babelex
blinker==1.4              # via flask-mail, flask-principal
click==7.0                # via flask
flask-babelex==0.9.4      # via flask-security
flask-login==0.5.0        # via flask-security
flask-mail==0.9.1         # via flask-security
flask-principal==0.4.0    # via flask-security
flask-security==3.0.0
flask-wtf==0.14.3         # via flask-security
flask==1.1.1
itsdangerous==1.1.0       # via flask, flask-security, flask-wtf
jinja2==2.11.1            # via flask, flask-babelex
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via jinja2, mako
passlib==1.7.2            # via flask-security
psycopg2-binary==2.8.4
python-dateutil==2.8.1    # via alembic
python-decouple==3.3
python-editor==1.0.4      # via alembic
pytz==2019.3              # via babel
six==1.14.0               # via python-dateutil
speaklater==1.3           # via flask-babelex
sqlalchemy-citext==1.5.post0
sqlalchemy==1.3.13
werkzeug==1.0.0           # via flask
wtforms==2.2.1            # via flask-wtf
